+++ .vscode/launch.json
@@ -1,0 +1,22 @@
+{
+    // Use IntelliSense to learn about possible attributes.
+    // Hover to view descriptions of existing attributes.
+    // For more information, visit: https://go.microsoft.com/fwlink/?linkid=830387
+    "version": "0.2.0",
+    "configurations": [
+        {
+            "type": "java",
+            "name": "Launch with Arguments Prompt",
+            "request": "launch",
+            "mainClass": "mathMagical",
+            "args": "'Club 7' 'Diamond 3' 'Diamond 4' 'Spade 10' 'Heart J' 'Spade 3' 'Spade 5'"
+        }
+        // {
+        //     "type": "java",
+        //     "name": "Launch mathMagical",
+        //     "request": "launch",
+        //     "mainClass": "mathMagical",
+        //     "projectName": "mathMagical_f23aa0a4"
+        // }
+    ]
+}
